+++ src/serializers/JsonSerializer.php
@@ -8,6 +8,7 @@
 namespace zhuravljov\yii\queue\serializers;
 
 use Yii;
+use yii\base\InvalidConfigException;
 use yii\base\Object;
 use yii\helpers\Json;
 
@@ -46,12 +47,16 @@ class JsonSerializer extends Object implements Serializer
     /**
      * @param mixed $data
      * @return array|mixed
+     * @throws InvalidConfigException
      */
     protected function toArray($data)
     {
         if (is_object($data)) {
             $result = [$this->classKey => get_class($data)];
             foreach (get_object_vars($data) as $property => $value) {
+                if ($property === $this->classKey) {
+                    throw new InvalidConfigException("Object cannot contain $this->classKey property.");
+                }
                 $result[$property] = $this->toArray($value);
             }
 
@@ -61,6 +66,9 @@ class JsonSerializer extends Object implements Serializer
         if (is_array($data)) {
             $result = [];
             foreach ($data as $key => $value) {
+                if ($key === $this->classKey) {
+                    throw new InvalidConfigException("Array cannot contain $this->classKey key.");
+                }
                 $result[$key] = $this->toArray($value);
             }