+++ tests/serializers/JsonSerializerTest.php
@@ -23,4 +23,14 @@ class JsonSerializerTest extends TestCase
     {
         return new JsonSerializer();
     }
+
+    /**
+     * @expectedException \yii\base\InvalidConfigException
+     */
+    public function testInvalidArrayKey()
+    {
+        $this->createSerializer()->serialize([
+            'class' => 'failed param',
+        ]);
+    }
 }